Create close
Research to see if we can open, read and write to the perl map and then close it afterwards.

At the moment, I don't think it is possible. This function may only be used to change the value of the file using a text editor. Not sure if it can read the map, use our function to find squares and replace empty (.) characters with full (x) characters, and then reprint the map with the changes.

So more research and testing needs to be done.